separated postgres event store from core functions
the core of the postgres event store are the functions building the sql statements.
that's the domain logic.

extracted postgres core functions into their own module.
core functions are static (pure functions).
