+++ src/eventstore/index.ts
@@ -1,3 +1,3 @@
-export { PostgresEventStore, IPostgresEventStoreOptions } from './postgres';
+export { PostgresEventStore, IPostgresEventStoreOptions } from './postgres-eventstore';
 export { EventFilter } from './filter';
 export * from './types';
